fix : searchBar 컴포넌트 분리 및 스타일 수정
