+++ src/components/searchBar/RecommendSearch.tsx
@@ -65,7 +65,7 @@ const RecommendSearch = ({
 
   return (
     <div className="fixed inset-0 z-200 flex items-start" onClick={onClose}>
-      <div className="absolute inset-0 ml-25 mt-16 bg-black/40" />
+      <div className="absolute inset-0 ml-25 mt-16 bg-black80/40" />
 
       <div
         className="relative w-[880px] max-w-[calc(100vw-400px)] min-w-[640px] bg-white rounded-[15px] shadow-lg ml-[347px] mt-16"
